+++ src/main/java/conspire/cards/colorless/GhostlyDefend.java
@@ -1,0 +1,66 @@
+package conspire.cards.colorless;
+
+import com.megacrit.cardcrawl.actions.common.GainBlockAction;
+import com.megacrit.cardcrawl.cards.AbstractCard;
+import com.megacrit.cardcrawl.characters.AbstractPlayer;
+import com.megacrit.cardcrawl.core.CardCrawlGame;
+import com.megacrit.cardcrawl.core.Settings;
+import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
+import com.megacrit.cardcrawl.localization.CardStrings;
+import com.megacrit.cardcrawl.monsters.AbstractMonster;
+
+import basemod.abstracts.CustomCard;
+import basemod.helpers.BaseModCardTags;
+import conspire.Conspire;
+
+public class GhostlyDefend extends CustomCard {
+    public static final String ID = "conspire:GhostlyDefend";
+    private static final CardStrings cardStrings = CardCrawlGame.languagePack.getCardStrings(ID);
+    public static final String NAME = cardStrings.NAME;
+    public static final String DESCRIPTION = cardStrings.DESCRIPTION;
+    private static final int COST = 1;
+    private static final int BLOCK_AMT = 5;
+
+    public GhostlyDefend() {
+        super(ID, NAME, Conspire.cardImage(ID), COST, DESCRIPTION, CardType.SKILL, CardColor.COLORLESS, CardRarity.BASIC, CardTarget.SELF);
+        this.baseBlock = BLOCK_AMT;
+        this.isEthereal = true;
+        this.tags.add(BaseModCardTags.BASIC_DEFEND);
+    }
+
+    public GhostlyDefend(AbstractCard card) {
+        this();
+        this.upgraded = card.upgraded;
+        this.timesUpgraded = card.timesUpgraded;
+        this.baseDamage = card.baseDamage;
+        this.baseBlock = card.baseBlock;
+        this.baseMagicNumber = card.baseMagicNumber;
+        this.cost = card.cost;
+        this.inBottleLightning = card.inBottleLightning;
+        this.inBottleFlame = card.inBottleFlame;
+        this.inBottleTornado = card.inBottleTornado;
+    }
+
+    @Override
+    public void use(AbstractPlayer p, AbstractMonster m) {
+        if (Settings.isDebug) {
+            AbstractDungeon.actionManager.addToBottom(new GainBlockAction(p, p, 50));
+        } else {
+            AbstractDungeon.actionManager.addToBottom(new GainBlockAction(p, p, this.block));
+        }
+    }
+
+    @Override
+    public AbstractCard makeCopy() {
+        return new GhostlyDefend();
+    }
+
+    @Override
+    public void upgrade() {
+        if (!this.upgraded) {
+            this.upgradeName();
+            this.upgradeBlock(3);
+        }
+    }
+}
+